Refactoring + first step to centralize provide object
