+++ dataland-frontend/tests/unit/TestUtils.ts
@@ -1,0 +1,12 @@
+import Keycloak, {KeycloakError, KeycloakPromise} from "keycloak-js";
+
+const a: string = "aaa"
+
+export function getInjectedKeycloakObjectsForTest(): any{
+    return {
+        getKeycloakInitPromise() {
+            return a;
+        },
+        keycloak_init: a,
+    }
+}
